+++ src/textual_enhanced/commands/__init__.py
@@ -3,10 +3,19 @@
 ##############################################################################
 # Local imports.
 from .command import Command
+from .common import CommonCommands, Help, Quit
 from .provider import CommandHit, CommandHits, CommandsProvider
 
 ##############################################################################
 # Exports.
-__all__ = ["CommandHit", "CommandHits", "Command", "CommandsProvider"]
+__all__ = [
+    "CommandHit",
+    "CommandHits",
+    "Command",
+    "CommandsProvider",
+    "CommonCommands",
+    "Help",
+    "Quit",
+]
 
 ### __init__.py ends here